refactor: configs APIファイル構造をRESTfulに整理
- index.ts: GET /api/v1/configs (一覧取得) を処理
- get.ts: 削除（将来のGET /api/v1/configs/:id用に予約）
- 重複コードを削除してよりクリーンな構造に
- TypeScript型定義を明示的に追加
